-Wall changes
git-svn-id: https://svn.r-project.org/R/branches/R-0-90-patches@7008 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/devPS.c
@@ -581,7 +581,7 @@ static void PostScriptWriteString(FILE *fp, char *str)
 	    fprintf(fp, "\\n");
 	    break;
 	case '-':
-	    if (isdigit(str[1]))
+	    if (isdigit((int)str[1]))
 		fputc(PS_minus, fp);
 	    else
 		fputc(*str, fp);
